raycast addition

+++ oengine/Utils.odin
@@ -24,8 +24,8 @@ Rad2Deg :: 180.0 / math.PI
 Color :: rl.Color
 char :: rune
 
-F32_MAX :: math.F32_MAX
-F32_MIN :: math.F32_MIN
+F32_MAX :: f32(math.F32_MAX)
+F32_MIN :: f32(math.F32_MIN)
 
 clr_to_arr :: proc(color: Color, $T: typeid) -> [4]T {
     return [4]T {
